requirements.txt
requests
pandas
validators
setuptools
configparser
google-cloud-storage
-e.